requirements.txt
python-telegram-bot==21.6
demucs==4.0.1
torch>=2.0.0
torchaudio>=2.0.0
ffmpeg-python==0.2.0
flask